+++ backend/auth_service/.dockerignore
