+++ backend/auth_service/src/schemas/templates_auth.py
