+++ backend/auth_service/src/scripts/passwords.py
@@ -1,0 +1,29 @@
+from datetime import datetime, timedelta, timezone
+
+from jose import jwt
+from passlib.context import CryptContext
+
+from backend.auth_service.config.database import get_auth_data
+
+
+def create_access_token(data: dict) -> str:
+    """Создает JWT токен"""
+    to_encode = data.copy()
+    expire = datetime.now(timezone.utc) + timedelta(days=30)
+    to_encode.update({"exp": expire})
+    auth_data = get_auth_data()
+    encode_jwt = jwt.encode(to_encode, auth_data["secret_key"], algorithm=auth_data["algorithm"])
+    return encode_jwt
+
+
+pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
+
+
+def get_password_hash(password: str) -> str:
+    """Хеширует пароль"""
+    return pwd_context.hash(password)
+
+
+def verify_password(plain_password: str, hashed_password: str) -> bool:
+    """Проверяет соответствие пароля хешу"""
+    return pwd_context.verify(plain_password, hashed_password)
