+++ backend/recsys_service/requirements_prod.txt
@@ -1,0 +1,11 @@
+pandas==2.2.2
+joblib==1.4.2
+pydantic==2.7.1
+pydantic-settings==2.2.1
+sentence-transformers==3.0.1
+uvicorn==0.30.1
+fastapi==0.111.0
+qdrant-client==1.10.1
+python-dotenv==1.0.1
+prometheus-fastapi-instrumentator==7.0.0
+requests==2.31.0
