--- backend/requirements_prod.txt
@@ -1,16 +1,0 @@
-pandas==2.2.2
-joblib==1.4.2
-pydantic==2.7.1
-pydantic-settings==2.2.1
-sentence-transformers==3.0.1
-uvicorn==0.30.1
-fastapi==0.111.0
-qdrant-client==1.10.1
-redis==5.0.7
-python-dotenv==1.0.1
-sqlalchemy==2.0.33
-asyncpg==0.29.0
-python-jose==3.3.0
-bcrypt ==4.0.1
-passlib==1.7.4
-prometheus-fastapi-instrumentator==7.0.0
