+++ backend/search_service/.dockerignore
@@ -1,0 +1,5 @@
+Dockerfile
+docker-compose.yml
+setup.py
+title_id_hash_map.pkl
+alembic
