+++ backend/search_service/src/exceptions/custom_exceptions.py
