+++ backend/search_service/src/models/trie.py
