+++ backend/search_service/src/schemas/tamplates_app.py
