--- backend/src/models/tables.py
@@ -1,46 +1,0 @@
-import os
-import uuid
-
-from dotenv import load_dotenv
-from sqlalchemy import Boolean, Column, String, Text
-from sqlalchemy.dialects.postgresql import UUID
-
-from backend.config.database import Base
-
-load_dotenv()
-
-
-class User(Base):
-    """Шаблон таблицы для пользователей"""
-
-    __tablename__ = f"{os.getenv('POSTGRE_USER_TABLE')}"
-
-    uid = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
-    login = Column(String)
-    password = Column(String)
-    is_user = Column(Boolean, default=True)
-    is_admin = Column(Boolean, default=False)
-    is_super_admin = Column(Boolean, default=False)
-
-    extend_existing = True
-
-    def __repr__(self):
-        return f"{self.__class__.__name__}(id={self.uid})"
-
-
-class Book(Base):
-    """Шаблон таблицы для книг"""
-
-    __tablename__ = f"{os.getenv('POSTGRE_BOOK_TABLE')}"  # Задаем имя таблицы в базе данных
-
-    ID = Column(String, primary_key=True)  # Первичный ключ
-    Title = Column(String)  # Строка с максимальной длиной 255 символов
-    Author = Column(String)  # Строка с максимальной длиной 255 символов
-    Link = Column(String)  # Строка с максимальной длиной 255 символов, может быть пустой
-    Image = Column(String)  # Строка с максимальной длиной 255 символов, может быть пустой
-    Category = Column(String)  # Строка с максимальной длиной 255 символов, может быть пустой
-    Description = Column(Text)  # Текст без ограничения длины
-    Info = Column(Text)  # Текст без ограничения длины
-
-    def __repr__(self):
-        return f"{self.__class__.__name__}(ID={self.ID}, Title={self.Title})"
